requirements.txt
aio-pika==3.0.1
aiofiles==0.4.0
aiohttp==3.4.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
click==6.7
coloredlogs==10.0
cytoolz==0.9.0.1
decorator==4.4.0
Django==2.2.2
django-ipware==2.1.0
django-mathfilters==0.4.0
django-smart-selects==1.5.4
earlgrey==0.0.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.2
eth-typing==2.1.0
eth-utils==1.6.0
funcsigs==1.0.2
future==0.16.0
geoip2==2.9.0
grpcio==1.19.0
grpcio-tools==1.19.0
gunicorn==19.9.0
httptools==0.0.13
humanfriendly==4.18
iconcommons==1.0.5.2
iconrpcserver==1.3.0
iconsdk==1.0.9
iconservice==1.3.0
idna==2.8
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.13.3
jsonrpcclient==2.6.0
jsonrpcserver==3.5.6
jsonschema==2.6.0
lazy-object-proxy==1.4.1
maxminddb==1.4.1
mccabe==0.6.1
msgpack==0.6.1
multidict==4.5.2
multipledispatch==0.5.0
mysqlclient==1.4.2.post1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pika==0.12.0
Pillow==6.0.0
plyvel==1.0.5
prompt-toolkit==1.0.16
protobuf==3.7.0
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.8.2
pygeoip==0.3.2
Pygments==2.4.2
pylint==2.3.1
pytz==2019.1
requests==2.22.0
sanic==18.12.0
Sanic-Cors==0.9.6
Sanic-Plugins-Framework==0.7.0
secp256k1==0.13.2
setproctitle==1.1.10
shortuuid==0.5.0
simplegeneric==0.8.1
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tbears==1.2.0
toolz==0.9.0
traitlets==4.3.2
typed-ast==1.4.0
ujson==1.35
urllib3==1.25.3
uvloop==0.12.2
wcwidth==0.1.7
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
